Refactoring of map related functions
